still failling one test case, but getting better
