+++ inc/raytracing.h
@@ -108,7 +108,6 @@ typedef struct s_quad
 	float	d;
 }	t_quad;
 
-
 //Material light properties
 typedef struct	s_material
 {
@@ -118,17 +117,6 @@ typedef struct	s_material
 	float	specular;
 	float	shininess;
 }	t_material;
-
-/**
- * @todo this is probably superfluous and redundant.
- */
-/*typedef struct s_sphere
-{
-	t_tuple		center;
-	float		radius;
-	float		**transform;
-	t_material	mat;
-}	t_sphere;*/
 
 typedef struct	s_light_vars
 {
@@ -220,7 +208,6 @@ void	matrix_fill_zero(t_matrix4 *m);
 //rays
 t_ray	create_ray(t_tuple vector, t_tuple point);
 t_tuple	position_ray(t_ray ray, float t);
-void	init_xs(t_xs *xs);
 t_i		hit(t_xs *xs, t_scene_obj *obj_from);
 t_ray	transform(t_ray r, t_matrix4 m);
 void	set_transform(t_scene_obj *obj, t_matrix4 m);
@@ -239,9 +226,12 @@ t_matrix4	view_transform(t_tuple from, t_tuple to, t_tuple up);
 void		init_camera(t_minirt *minirt);
 t_ray		ray_for_pixel(t_minirt *minirt, t_camera *c, int px, int py);
 
-//intersections
+//intersections.c
 t_xs		*intersects_sphere(t_scene_obj *obj, t_ray r, t_xs *xs);
 t_xs		*intersect(t_minirt *minirt, t_scene_obj *obj, t_ray ray, t_xs *xs);
+
+//intersect_cylinder.c
+t_xs	*intersects_cylinder(t_scene_obj *obj, t_ray r, t_xs *xs);
 
 //transformation functions
 t_matrix4	cylinder_rotation(t_scene_obj *obj);
